+++ client/src/components/Brute/WeaponTooltip.tsx
@@ -7,6 +7,7 @@ import StatColor from '../../utils/StatColor';
 import { useAuth } from '../../hooks/useAuth';
 import { TierStar } from './TierStar';
 import { FightModifier } from '@labrute/prisma';
+import TieredStat from '../TieredStat';
 
 const textShadowBase = '-1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, 1px 1px 0 #000';
 const textProps = {
@@ -50,8 +51,6 @@ const WeaponTooltip = ({
 
     const label = stat === 'tempo' ? 'hitSpeed' : stat === 'toss' ? 'drawChance' : stat;
     const color = stat === 'tempo' || stat === 'toss' ? { opacity: 0.7 } : { color: StatColor[stat] };
-    const uniqueValue = !tieredWeapon
-      || tieredWeapon[stat].every((v) => v === tieredWeapon[stat]?.[0]);
 
     const formatter = tierFormatter ?? ((currentTier: number) => {
       // Those stats will always use a formatter
@@ -70,21 +69,14 @@ const WeaponTooltip = ({
           {t(label)}:
           {' '}
           <Text component="span" bold sx={color} {...textProps}>
-            {uniqueValue ? formatter(tier) : (
-              <>
-                [
-                {tieredWeapon[stat].map((_, index) => (
-                  // eslint-disable-next-line react/no-array-index-key
-                  <Fragment key={index}>
-                    <Box sx={{ fontWeight: index === tier - 1 ? 'bold' : 'normal' }} component="span">
-                      {formatter(index + 1)}
-                    </Box>
-                    {index < (tieredWeapon[stat]?.length ?? 0) - 1 ? '/' : ''}
-                  </Fragment>
-                ))}
-                ]
-              </>
-            )}{percent ? '%' : ''}
+            {tieredWeapon ? (
+              <TieredStat
+                values={tieredWeapon[stat]}
+                tier={tier}
+                formatter={(_, index) => formatter(index + 1)}
+              />
+            ) : formatter(tier)}
+            {percent ? '%' : ''}
           </Text>
         </Text>
       );
@@ -95,23 +87,22 @@ const WeaponTooltip = ({
 
     return (
       <Text bold sx={{ color: StatColor[stat], textShadow }} {...textProps}>
-        {(tieredWeapon ? (tieredWeapon[stat][tieredWeapon.tier - 1] ?? 0) : BASE_FIGHTER_STATS[stat]) > 0 ? '+' : '-'}
-        {uniqueValue ? Math.abs(formatter(tier)) : (
+        {tieredWeapon ? (
           <>
-            [
-            {tieredWeapon[stat].map((_, index) => (
-              // eslint-disable-next-line react/no-array-index-key
-              <Fragment key={index}>
-                <Box sx={{ fontWeight: index === tier - 1 ? 'bold' : 'normal' }} component="span">
-                  {Math.abs(formatter(index + 1))}
-                </Box>
-                {index < (tieredWeapon[stat]?.length ?? 0) - 1 ? '/' : ''}
-              </Fragment>
-            ))}
-            ]
+            {(tieredWeapon[stat][tieredWeapon.tier - 1] ?? 0) > 0 ? '+' : '-'}
+            <TieredStat
+              values={tieredWeapon[stat]}
+              tier={tier}
+              formatter={(_, index) => Math.abs(formatter(index + 1))}
+            />
+          </>
+        ) : (
+          <>
+            {BASE_FIGHTER_STATS[stat] > 0 ? '+' : '-'}
+            {Math.abs(formatter(tier))}
           </>
         )}
-        {percent ? '%' : ''} {t(label)}
+        {percent ? '%' : ''}{' '}{t(label)}
       </Text>
     );
   };